requirements.txt
beautifulsoup4==4.14.3
fastapi==0.135.1
google_api_python_client==2.192.0
google_auth_oauthlib==1.3.0
imap_tools==1.11.1
pandas==3.0.1
protobuf==7.34.0
pydantic==2.12.5
PyJWT==2.12.1
pyparsing==3.3.2
pytest==9.0.2
python-dotenv==1.2.2
python_dateutil==2.9.0.post0
Requests==2.32.5
rich==14.3.3
SQLAlchemy==2.0.48
supabase==2.28.2
urllib3==2.6.3
uvicorn==0.42.0
pyinstaller==6.15.0
